requirements.txt
tqdm==4.51.0
numpy==1.18.5
torch==1.8.1
torchvision==0.2.2
Pillow==8.2.0
